Add support for masking prefix/suffix patterns
This is a slight generalization of the model used by Gitlab in their
token prefix matching. The prefix is a fixed string which must be
matched in full, and the suffix is an alphabet which is only matched
once the prefix has matched.

For example, (prefix=abc, suffix=def) matches abcfesdfedf and abc but
not abd.

As this prefix suffix model is a strict subset of the previously
supported masking, the old type of match is simply converted into a
prefix with an empty suffix. Note that whether to mask the prefix is
selectable - so a former input like "string" becomes (prefix="string",
suffix="", mask_prefix=true).

There are two main subtleties in this implementation:

- When there are many (prefix,suffix) pairs, and the prefixes match
  the suffixes, we get a state space explosion. There is a control for
  this case for some types of input, in that we calculate at what point
  a match of one pair is ignorable, because it is entirely covered by an
  existing match. This collapses the number of states in the common case.

- When matching suffixes, there is a state explosion when we attempt
  to enumerate all the possible suffixes. But this is unnecessary,
  because the actual content matched to the suffix is irrelevant
  (suffixes are always masked) and this allows us to simply store an
  offset per state (as text_offset) in the text. This in turn brings
  down the number of states dramatically.